backend/requirements.txt
# FastAPI and Server (Essential only)
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.0

# Database (PostgreSQL with SQLAlchemy)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
